api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.39
asyncpg==0.30.0
alembic==1.15.2
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.1
aioredis==2.0.1
httpx==0.28.1
slowapi==0.1.9
pytest==8.3.5
pytest-asyncio==0.25.3
aiosqlite==0.21.0
